:recycle: BaseViewModel Refactoring
